# docs/swagger.yaml
basePath: /
definitions:
  controllers.AssignTaskRequest:
    properties:
      assign_to_email:
        type: string
      description:
        type: string
      due_date:
        example: yyyy-mm-dd
        type: string
      priority:
        type: string
      reminder_at:
        example: "2025-01-02T10:00:00+05:30"
        type: string
      status:
        type: string
      task_group:
        type: string
      title:
        type: string
    required:
    - assign_to_email
    - title
    type: object
  controllers.CreateTaskRequest:
    properties:
      description:
        type: string
      due_date:
        description: user passes string in YYYY-MM-DD
        example: yyyy-mm-dd
        type: string
      priority:
        enum:
        - medium
        - high
        - low
        type: string
      reminder_at:
        example: "2025-01-02T10:00:00+05:30"
        type: string
      status:
        enum:
        - pending
        - in_progress
        - completed
        type: string
      task_group:
        enum:
        - personal
        - office
        - shopping
        - family
        - friends
        - education
        - health
        - travel
        - food
        type: string
      title:
        maxLength: 100
        type: string
    required:
    - title
    type: object
  controllers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 4
        type: string
      username:
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  controllers.UpdateTaskRequest:
    properties:
      description:
        type: string
      due_date:
        description: string from client
        example: yyyy-mm-dd
        type: string
      priority:
        enum:
        - medium
        - high
        - low
        type: string
      status:
        description: optional
        enum:
        - pending
        - in_progress
        - completed
        type: string
      task_group:
        enum:
        - personal
        - office
        - shopping
        - family
        - friends
        - education
        - health
        - travel
        - food
        type: string
      title:
        maxLength: 100
        type: string
    required:
    - description
    - due_date
    - task_group
    - title
    type: object
  models.Task:
    properties:
      assigned_by:
        type: string
      assigned_by_id:
        type: integer
      assigned_role:
        description: admin / user
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      is_overdue:
        type: boolean
      priority:
        type: string
      reminder_at:
        type: string
      status:
        type: string
      task_group:
        enum:
        - personal
        - office
        - shopping
        - family
        - friends
        - education
        - health
        - travel
        - food
        example: personal
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      role:
        description: user | admin
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simple Todo API with Gin, GORM, MySQL, Swagger
  title: Todo API
  version: "1.0"
paths:
  /admin/searchTask:
    get:
      consumes:
      - application/json
      description: Admin-only API to search tasks by title or description
      parameters:
      - description: Search keyword (matches title or description)
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of matching tasks with count
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Search query cannot be empty
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized (missing or invalid token)
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (admin access required)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Search tasks (Admin only)
      tags:
      - Admin
  /admin/tasks:
    get:
      description: Admin can view all tasks across all users
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.Task'
              type: array
            type: object
        "403":
          description: Admin access only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch tasks
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all tasks (Admin only)
      tags:
      - Admin
  /admin/tasksByFilter:
    get:
      description: Retrieve all tasks with optional filters (priority, status, due_date)
        (JWT required)
      parameters:
      - description: Filter by priority (high, medium, low)
        in: query
        name: priority
        type: string
      - description: Filter by status (pending, in-progress, completed)
        in: query
        name: status
        type: string
      - description: Filter by due date (YYYY-MM-DD)
        in: query
        name: due_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Invalid due_date format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch tasks
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all tasks  (Admin only)
      tags:
      - Admin
  /admin/users:
    get:
      description: Admin can view the list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.User'
              type: array
            type: object
        "403":
          description: Admin access only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch users
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all users (Admin only)
      tags:
      - Admin
  /deleteTask/{id}:
    delete:
      description: Delete a task by its id (JWT required)
      parameters:
      - description: Task id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Task deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid task id
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete task
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - tasks
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid username or password
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input or username already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error (failed to hash password)
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /reports/overdue:
    get:
      consumes:
      - application/json
      description: Returns all overdue (past due date) tasks for the currently logged-in
        user.
      produces:
      - application/json
      responses:
        "200":
          description: List of overdue tasks
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user's overdue tasks
      tags:
      - Reports
  /reports/tomorrow:
    get:
      consumes:
      - application/json
      description: Returns all tasks whose due date is exactly tomorrow for the logged-in
        user.
      produces:
      - application/json
      responses:
        "200":
          description: List of tomorrow's tasks
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user's tasks due tomorrow
      tags:
      - Reports
  /reports/week:
    get:
      consumes:
      - application/json
      description: Returns all tasks due within the next 7 days for the logged-in
        user.
      produces:
      - application/json
      responses:
        "200":
          description: List of next 7 days tasks
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user's tasks due in next 7 days
      tags:
      - Reports
  /task:
    post:
      consumes:
      - application/json
      description: Add a new task (JWT required)
      parameters:
      - description: Task input
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - tasks
  /tasks:
    get:
      description: Retrieve all tasks from database (no filter) (JWT required)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
      security:
      - BearerAuth: []
      summary: Get all tasks
      tags:
      - tasks
  /tasks/{id}:
    get:
      description: Retrieve a single task by its ID (JWT required)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a task by ID
      tags:
      - tasks
  /tasks/assign:
    post:
      consumes:
      - application/json
      description: Admin can assign a task to any user. A normal user can assign a
        task only to another user (not admin).
      parameters:
      - description: Task assignment payload
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/controllers.AssignTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Task assigned successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Assignee not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign a task to a user
      tags:
      - tasks
  /tasks/group/{group}:
    get:
      description: Get tasks filtered by specific groups like personal, education,
        office, etc.
      parameters:
      - description: Task Group
        enum:
        - personal
        - office
        - shopping
        - family
        - friends
        - education
        - health
        - travel
        - food
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
      security:
      - BearerAuth: []
      summary: Get tasks by group
      tags:
      - tasks
  /tasksByFilter:
    get:
      description: Retrieve all tasks with optional filters (priority, status, due_date)
        (JWT required)
      parameters:
      - description: Filter by priority (high, medium, low)
        in: query
        name: priority
        type: string
      - description: Filter by status (pending, in-progress, completed)
        in: query
        name: status
        type: string
      - description: Filter by due date (YYYY-MM-DD)
        in: query
        name: due_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Invalid due_date format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch tasks
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all tasks
      tags:
      - tasks
  /updateTasks/{id}:
    put:
      consumes:
      - application/json
      description: Replace all fields of a task by its ID. (JWT required)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Task input
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update task
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - tasks
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"